Core:
Rollback parts of last two patches,
* Brought back sqllite (temp)
* Updated ZThreads fo more future tests
* Reverted the build updates for know.
-  Will update them back to what they were
-  in a later revision.
-  this also pulls PCH back out of the build,
-  no big issue as of right now.

+++ dep/src/zthread/posix/FtimeStrategy.h
@@ -52,11 +52,11 @@ public:
   }
 
   inline unsigned long seconds() const {
-    return (unsigned long)_value.time;
+    return _value.time;
   }
 
-  inline unsigned long milliseconds() const {
-    return _value.millitm;
+  inline unsigned long milliseconds() const {  
+    return _value.millitm;    
   }
 
   unsigned long seconds(unsigned long s) {
@@ -82,4 +82,3 @@ public:
 };
 
 #endif // __ZTTIMESTRATEGY_H__
-
